pyproject.toml

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "metric_transpi"
authors = [
    {name = "Benjamin Joyen", email = "benjamin.joyen@decathlon.com"},
]
description = "Convert metric YAML into Tableau Pulse payload"
requires-python = ">=3.8"
dynamic = ["version"]
dependencies = [
    "pyyaml",
    "jinja2",
    "requests",
    "rich-click",
    "urllib3 >= 1.25.3, < 3.0.0",
    "python-dateutil >= 2.8.2",
    "pydantic >= 2", 
    "python-dotenv"
]

[tool.setuptools.packages]
find = {} 

[project.scripts]
metric-cli = "metric_transpi:cli.cli"

[tool.pytest.ini_options]
addopts = "--cov=metric_transpi"

[pytest]
pythonpath = "."